modify user-api & modify res.json to unified format
